+++ backend/nodes/urls.py
@@ -8,6 +8,7 @@ app_name = "nodes"
 router = routers.get_router()
 
 router.register("nodes/nodes", views.NodeModelViewSet, basename="nodes")
+router.register("nodes/methods", views.MethodNodeModelViewSet, basename="methods")
 router.register("nodes/spaces", views.SpaceModelViewSet, basename="spaces")
 router.register("nodes/versions", views.DocumentVersionModelViewSet, basename="document-versions")